Implemented SimpleGUI.
Allows the usage of buttons, text boxes, check boxes, and others. Provides the framework in the graphics loop to support them. Added tests which interactively test a small subset of the functionality.

Modified the AllInspections set to be more useful instead of ignoring them all.

Updated to JDK1.8, but still using language level JDK1.7.

Changed the default logging level to WARNING. Since WARNINGs and above are only generated when something goes wrong, the log should be clean if nothing goes wrong.

Moved the PerformanceTest from being a unit test to being an interactive user test, same as the GUI tests.

Modified the currently used config file so that the program fills more of a 1080p monitor.